+++ cli/src/commands/bookmark/mod.rs
@@ -12,6 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+mod advance;
 mod create;
 mod delete;
 mod forget;
@@ -37,6 +38,8 @@ use jj_lib::str_util::StringExpression;
 use jj_lib::str_util::StringMatcher;
 use jj_lib::view::View;
 
+use self::advance::BookmarkAdvanceArgs;
+use self::advance::cmd_bookmark_advance;
 use self::create::BookmarkCreateArgs;
 use self::create::cmd_bookmark_create;
 use self::delete::BookmarkDeleteArgs;
@@ -47,6 +50,7 @@ use self::list::BookmarkListArgs;
 use self::list::cmd_bookmark_list;
 use self::r#move::BookmarkMoveArgs;
 use self::r#move::cmd_bookmark_move;
+use self::r#move::move_bookmarks_with_args;
 use self::rename::BookmarkRenameArgs;
 use self::rename::cmd_bookmark_rename;
 use self::set::BookmarkSetArgs;
@@ -71,6 +75,8 @@ use crate::ui::Ui;
 ///     https://docs.jj-vcs.dev/latest/bookmarks
 #[derive(clap::Subcommand, Clone, Debug)]
 pub enum BookmarkCommand {
+    #[command(visible_alias("a"))]
+    Advance(BookmarkAdvanceArgs),
     #[command(visible_alias("c"))]
     Create(BookmarkCreateArgs),
     #[command(visible_alias("d"))]
@@ -96,6 +102,7 @@ pub fn cmd_bookmark(
     subcommand: &BookmarkCommand,
 ) -> Result<(), CommandError> {
     match subcommand {
+        BookmarkCommand::Advance(args) => cmd_bookmark_advance(ui, command, args),
         BookmarkCommand::Create(args) => cmd_bookmark_create(ui, command, args),
         BookmarkCommand::Delete(args) => cmd_bookmark_delete(ui, command, args),
         BookmarkCommand::Forget(args) => cmd_bookmark_forget(ui, command, args),